June 2025 map versions
These were the maps we used in our initial 2 tournaments. 

The maps named for the towns were played in the first two rounds of the tournament, representing beach invasion and defense, respectively. The "Airport" and "Railway" maps were used in the final matches of the tournament, representing later battles for key points in the Normandy infrastructure, the airfield near Carpiquet and the Caen-Bayeux railway station near Bretteville-l'Orguilleuse.